apicostservice 可用了

+++ CoreLMS.Core/DataTransferObjects/ProjectCostViewModel.cs
@@ -8,7 +8,7 @@
 
         public string YearDate { get; set; }
 
-        public int TotalAmount { get; set; }
+        public long TotalAmount { get; set; }
 
     }
 }
